Move sale details dialog paths to qs
