+++ saleor/static/dashboard-next/discounts/index.tsx
@@ -10,14 +10,13 @@ import {
   saleAddPath,
   saleListPath,
   salePath,
+  SaleUrlQueryParams,
   voucherAddPath,
   voucherListPath,
   voucherPath
 } from "./urls";
 import SaleCreateView from "./views/SaleCreate";
-import SaleDetailsViewComponent, {
-  SaleDetailsQueryParams
-} from "./views/SaleDetails";
+import SaleDetailsViewComponent from "./views/SaleDetails";
 import SaleListViewComponent, { SaleListQueryParams } from "./views/SaleList";
 import VoucherCreateView from "./views/VoucherCreate";
 import VoucherDetailsViewComponent, {
@@ -41,11 +40,10 @@ const SaleListView: React.StatelessComponent<RouteComponentProps<{}>> = ({
 const SaleDetailsView: React.StatelessComponent<
   RouteComponentProps<{ id: string }>
 > = ({ match, location }) => {
-  const { after, before, tab } = parseQs(location.search.substr(1));
-  const params: SaleDetailsQueryParams = {
-    after,
-    before,
-    tab: saleDetailsPageTab(tab)
+  const { activeTab, ...qs } = parseQs(location.search.substr(1));
+  const params: SaleUrlQueryParams = {
+    ...qs,
+    activeTab: saleDetailsPageTab(activeTab)
   };
   return (
     <SaleDetailsViewComponent
